youtube-mcp/pyproject.toml
[project]
name = "youtube-mcp"
version = "0.1.0"
description = "YouTube MCP Server for Claude Code - Upload, manage, and analyze YouTube videos"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.0.0",
    "google-api-python-client>=2.100.0",
    "google-auth-oauthlib>=1.1.0",
    "google-auth-httplib2>=0.2.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[project.scripts]
youtube-mcp = "youtube_mcp.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/youtube_mcp"]

[tool.hatch.build.targets.sdist]
include = ["src/youtube_mcp"]
